requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==3.2.20
django-allauth==0.54.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kombu==5.3.6
oauthlib==3.2.2
Pillow==10.0.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
